feat(cli): generation of test entries
This wires in an alpha level binary that can generate
skeleton "code-blog" entries.  These entries are hard
coded to have a persistence root of:

> `/tmp/mdbook-journal-test-cli/`

The skeleton topic has a single required meta-data key
of `title`.  You can play with watching it fail if you
submit an empty value.

Lastly, the full path is printed to standard out of
where the demo journal entry was saved to.  There is
clean up of these and will be on you to remember this.