feat: Simple Directory Navigation Layout 📂
**This marks the first official release of the journal system!**

By default now every topic is appended to the end of the
navigation and presents a structure similar to the file
system in which they are saved.

The goal will be to support custom templating for every folder
index; however, this is still useful enough now that I am ready
to begin using it for myself.