fix: Processor Honors Context Configuration
The mdBook processor flow was not honoring the configuration
being passed via STDIN.  This would manifest as errors in
certain build-flow configurations.  What was specifically
happening is if the mdbook process was being run from a remote
location that did not align with the book's base directory
it would not scan the correct data.