fix: Empty Topics and Assets No Longer Crash Journal
When a topic is defined but did not yet have an entry
all of the journal functionality would not work.  This
corrects that bug and reports no entries for a topic.

It was pointed out that supplemental assets placed among
the entries in the journal would crash it.  This was because
it was not correctly excluding invalid files and has been
corrected.  Please feel free to keep images in the same
directory as your markdown files.